GENIO: media: platform: mtk-mdp3: prevent overflow in calc_afbc_block
Added a cast to u64 for *block_x in the multiplication to prevent potential
overflow when calculating header_sz.

Coverity Report:
calc_afbc_block:
(1) Event overflow_before_widen:
    Potentially overflowing expression "*block_x * block_y" with type
    "unsigned int" is evaluated using 32-bit arithmetic.

CR-Id: AUTO00796894

+++ drivers/media/platform/mediatek/mdp3/mdp_alg_wrot.c
@@ -453,7 +453,7 @@ static void calc_afbc_block(u32 bits_per_pixel, u32 y_stride, u32 height,
 
 	*block_x = ((y_stride << 3) / bits_per_pixel + 31) >> 5;
 	block_y = (WROT_AFBC_ALIGN(height) + 7) >> 3;
-	header_sz = ((((*block_x * block_y) << 4) + 1023) >> 10) << 10;
+	header_sz = ((((u64)(*block_x) * block_y) << 4) + 1023) >> 10 << 10;
 
 	*addr_c = iova[0];
 	*addr_v = iova[2];
